Move FFI related things in one place.
This also factors things specific to the C calling
convention into a separate module.
