+++ cryptol.cabal
@@ -206,11 +206,7 @@ library
                        Cryptol.Backend,
                        Cryptol.Backend.Arch,
                        Cryptol.Backend.Concrete,
-                       Cryptol.Backend.FFI,
-                       Cryptol.Backend.FFI.Error,
-                       Cryptol.Backend.FFI.Call,
-                       Cryptol.Backend.FFI.ValExport,
-                       Cryptol.Backend.FFI.ValImport,
+                       
                        Cryptol.Backend.FloatHelpers,
                        Cryptol.Backend.Monad,
                        Cryptol.Backend.SeqMap,
@@ -222,6 +218,13 @@ library
                        Cryptol.Eval.Concrete,
                        Cryptol.Eval.Env,
                        Cryptol.Eval.FFI,
+                       Cryptol.Eval.FFI.ForeignSrc,
+                       Cryptol.Eval.FFI.Error,
+                       Cryptol.Eval.FFI.C,
+                       Cryptol.Eval.FFI.Abstract,
+                       Cryptol.Eval.FFI.Abstract.Call,
+                       Cryptol.Eval.FFI.Abstract.Export,
+                       Cryptol.Eval.FFI.Abstract.Import,
                        Cryptol.Eval.FFI.GenHeader,
                        Cryptol.Eval.Generic,
                        Cryptol.Eval.Prims,
