+++ src/Cryptol/Eval/FFI/Abstract.hs
@@ -1,0 +1,5 @@
+-- | How to call foreign function, following the "abstract" calling convention
+module Cryptol.Eval.FFI.Abstract where
+
+
+
